Fix usage of the Youtube API
