+++ requirements.lock
@@ -4,6 +4,7 @@ args==0.1.0
 async-timeout==1.2.1
 boto==2.48.0
 certifi==2017.4.17
+characteristic==14.3.0
 chardet==3.0.4
 click==6.7
 clint==0.5.1
@@ -22,6 +23,7 @@ pyasn1==0.2.3
 pyasn1-modules==0.0.9
 pydub==0.19.0
 pyquery==1.2.17
+python-dateutil==2.6.1
 requests==2.18.1
 rsa==3.4.2
 sh==1.12.14
